combine selection and bible text into the same componenet
