--- src/BibleSelect.css
@@ -1,6 +1,0 @@
-
-
-.book-select select {
-    margin: 0 5px;
-    border: dotted 313131 1px;
-}
