+++ src/ReadingPane.component.js
@@ -14,6 +14,8 @@ class ReadingPane extends React.Component {
   }
 
   state = {
+    selectedbookId: this.props.bookId,
+    selectedChapter: this.props.chapter,
     selectedVerse: 1,
     selectedVerseTarget: null,
   }
@@ -24,6 +26,22 @@ class ReadingPane extends React.Component {
     for (let verse of verses) {
       verse.addEventListener("click", this.handleSelectVerseEvent);
     }
+  }
+
+  handleBookSelection = (event) => {
+    this.setState({
+        selectedbookId: event.target.value,
+        selectedChapter: 1
+    })
+    this.props.changeBookChapterRequest(event.target.value, 1);
+}
+
+  handleChapterSelection = (event) => {
+      this.setState({
+          selectedChapter: event.target.value
+      })
+      // call parent function to change selected
+      this.props.changeBookChapterRequest(this.state.selectedbookId, parseInt(event.target.value));
   }
 
   // handle verse selected event
@@ -53,24 +71,40 @@ class ReadingPane extends React.Component {
       return <div></div>
     }
     const verses = this.props.data['verses'];
+    let menu = [];
+    for (let i = 1; i <= this.props.bibleIndex[this.state.selectedbookId].chapters; i++) {
+        menu.push(<option value={i} key={i}>{i}</option>)
+    }
 
     return (
       <div className="reading-pane">
-        <div className="chapter">
-          <div className="title">{this.props.bibleIndex[this.props.bookId].title} {this.props.chapter}</div>
-          {verses.map( verse => (
-            <span className="verse" 
-              key={`${this.props.bookId}.${verse.chapter}.${verse.verse}`}
-              data-book={this.props.bookId}
-              data-chapter={this.props.chapter}
-              data-verse={verse.verse}
-              onClick={this.handleSelectVerseEvent} >
-              <span className="label">{verse.verse}</span>
-              <span className="content">
-                {verse.text}
+        <div className="book-select">
+          <select className="books-select" value={this.state.selectedbookId} onChange={this.handleBookSelection}>
+              {Object.keys(this.props.bibleIndex).map(bookId => (
+                  (<option value={bookId} key={bookId}>{this.props.bibleIndex[bookId].title}</option>)
+              ))}
+          </select>
+          <select className="chapters-select" value={this.state.selectedChapter} onChange={this.handleChapterSelection}>
+              {menu}
+          </select>
+        </div>
+        <div className="content-pane">
+          <div className="chapter">
+            <div className="title">{this.props.bibleIndex[this.props.bookId].title} {this.props.chapter}</div>
+            {verses.map( verse => (
+              <span className="verse" 
+                key={`${this.props.bookId}.${verse.chapter}.${verse.verse}`}
+                data-book={this.props.bookId}
+                data-chapter={this.props.chapter}
+                data-verse={verse.verse}
+                onClick={this.handleSelectVerseEvent} >
+                <span className="label">{verse.verse}</span>
+                <span className="content">
+                  {verse.text}
+                </span>
               </span>
-            </span>
-          ))}
+            ))}
+          </div>
         </div>
       </div>
     );
